Added http error handling for err no 400

+++ backend/app.py
@@ -1,12 +1,9 @@
 import os
-from flask import Flask, jsonify, Response, request
+from flask import Flask, jsonify, Response, request, abort
 from db_manager import DatabaseManager
 
 app = Flask(__name__)
 
-
-# 400
-#
 
 # POST endpoint
 
@@ -14,9 +11,7 @@ app = Flask(__name__)
 # GET endpoint
 @app.route('/users')
 def get_users():
-    db = DatabaseManager(database=os.getenv('database'),
-                         user=os.getenv('user'),
-                         password=os.getenv('password'),
+    db = DatabaseManager(database=os.getenv('database'), user=os.getenv('user'), password=os.getenv('password'),
                          host=os.getenv('host'))
     all_users = db.get_users()
     db.close_connection()
@@ -29,18 +24,18 @@ def get_users():
 # DELETE endpoint
 @app.route('/users', methods=['DELETE'])
 def delete_user():
-    db = DatabaseManager(database=os.getenv('database'),
-                         user=os.getenv('user'),
-                         password=os.getenv('password'),
-                         host=os.getenv('host'))
-    try:
-        discord_id = request.form.get('discord')
+    discord_id = request.form.get('discord')
+    if discord_id:  # tutaj spr czy podany id w formularzu - jak tak, to idzie dalej
+        db = DatabaseManager(database=os.getenv('database'),
+                             user=os.getenv('user'),
+                             password=os.getenv('password'),
+                             host=os.getenv('host'))
         result = db.delete_user(discord_id)
-        return Response(result, mimetype='application/json'), 200
-    except KeyError:
-        return jsonify({"error": "User not found or incorrect discord id"})
-    finally:
         db.close_connection()
+        return Response(result, mimetype='application/json')
+    else:  # tutaj wyrzuca błąd i przerywa operację, jeśli brak id
+        abort(400,
+              description='Invalid request. Required data missing (discord id)')
 
 
 if __name__ == '__main__':
